fixed ui; added sort by hibernate, count by hibernate;

+++ model/src/main/java/com/senla/hotel/controller/ResidentController.java
@@ -25,7 +25,7 @@ public class ResidentController {
         return residentService.showResidents(sortField);
     }
 
-    public int showCountResidents() throws PersistException {
+    public Long showCountResidents() throws PersistException {
         return residentService.showCountResidents();
     }
